politique et autorisation en cours d'implementation pour les parametres

+++ app/Http/Controllers/Caisse/CaissierController.php
@@ -19,6 +19,7 @@ class CaissierController extends Controller
 
     public function store(Request $request): JsonResponse
     {
+        $this->authorize('create', Caissier::class);
         $request->validate(Caissier::RULES);
         $caissier = new Caissier($request->all());
         $caissier->codeGenerate();
@@ -30,22 +31,24 @@ class CaissierController extends Controller
     public function trash(int $id): JsonResponse
     {
         $caissier = Caissier::findOrFail($id);
+        $this->authorize('delete', $caissier);
         $caissier->delete();
         $message = "Le caissier: $caissier->code a été supprimé avec succès.";
-        // TODO: lorsque le caissier est supprimé il faut attribuer forcement ces bordereaux à une autre sinon on empêche la suppression
         return response()->json(['message' => $message]);
     }
 
     public function show(int $id): JsonResponse
     {
         $caissier = Caissier::with('attributions')->findOrFail($id);
+        $this->authorize('view', $caissier);
         return response()->json(['caissier' => $caissier]);
     }
 
     public function attribuate(Request $request): JsonResponse
     {
-        $request->validate(Caissier::ATTRIBUTION_RULES);
         $caissier = Caissier::findOrFail($request->caissier_id);
+        $this->authorize('view', $caissier);
+        $request->validate(Caissier::ATTRIBUTION_RULES);
         foreach ($request->dates as $dates) {
             $caissier->attributions()->attach($request->guichet_id, ['date' => $dates]);
         }
